Prefs.cpp: fix usage of -w and -W options
The -w and -W options allow to specify the HTTP/HTTPS ports the
web interface server will be listening to. By prepending a colon
to the port, it is possible to bind the web interface to the loopback
interface so that it can be accessed only by localhost. This commit
adds the latter feature to ntopng's usage message.

git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntopng@8638 a02cd0c1-8e76-42e1-a119-56f9641475e2