as_data: implement throughput trend in AS view
This commit adds to the AS view the indication about the current
throughput trend of each AS, as it is already done for hosts. The
new functionality is implemented by keeping, for each host, also
the difference between the last sampling of the throughput and the
current sampling, so that the global difference can be quickly
retrieved and grouped by AS by any Lua script.

git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntopng@8659 a02cd0c1-8e76-42e1-a119-56f9641475e2