as_stats: fix per-AS view automatic refresh
This commit actually implements the automatic refresh of the per-AS
view by also fixing the related javscript functions. The refresh is
performed every 10 seconds as it already happens for the hosts view.
A quite big portion of the get_as_data Lua script is refactored so
that the script can be used to generate both the JSON for the whole
AS list and the JSON for a single AS. The commit also fixes a type
mismatch for the AS identifier, which was a string for a real AS
name and an integer in case the AS was private or not identified
for some reason.

git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntopng@8660 a02cd0c1-8e76-42e1-a119-56f9641475e2