scripts/lua: add per-VLAN view in Hosts menu
This commit adds an item to the Hosts menu which allows to view
active IPs by VLAN. A huge portion of the code hosted by the
get_as_data Lua script is reworked to allow grouping by column.

This commit leaves one open TODO about the name of VLANs to be
viewed in the interface table.

git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntopng@8663 a02cd0c1-8e76-42e1-a119-56f9641475e2