top_talkers.lua: refactor getHistoricalTop*() function
This commit factors out common code from the getHistoricalTopTalkers()
and getHistoricalTopASs() functions into a common functions called
with a single parameter of the filename to be retrieved. Also a part
of the filename building process could be factored out, but for clarity
that part of code is still left as is.

git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntopng@8670 a02cd0c1-8e76-42e1-a119-56f9641475e2