if_stats: add Top Minute VLANs to historical interface view
This commit adds to the Historical tab of the interface details page
information about the top talker VLANs. This also makes the getTopASs()
function generic so that it is able to get top hosts information grouped
by any column. The name of the VLAN as shown in the page links to details
about active hosts pertaining to such VLAN.

git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntopng@8673 a02cd0c1-8e76-42e1-a119-56f9641475e2