Stats: move stats management to new class and top minute stats to a DB
This commit implements a new class handling storage of top minute stats.
The class exposes an abstraction layer that allows callers to store
stats at a sampling point identified by the current minute or to
retrieve stats of a previously-inserted sampling point. Private methods
of the class allow to store values on a DB or on the filesystem as it
currently happens.
This commit also changes storage of top minute stats so that it is
performed on a per-interface database.

git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntopng@8705 a02cd0c1-8e76-42e1-a119-56f9641475e2