if_stats.lua: fix sorting in Protocols interface stats
Currently, the Protocols interface stats are not being sorted
correctly. This commit lets new rows be inserted and not appended
during an update and signals to the TableSorter plugins that an
update has been performed whenever it is. It also removes a
timeout on the cache update of the TableSorter plugins. The commit
also lets TableSorter correctly parse network throughput cells
content and sort the content accordingly.

git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntopng@8714 a02cd0c1-8e76-42e1-a119-56f9641475e2