StatsManager: change interface by adding cache name parameter
This commit adds an extra parameter to the StatsManager interface,
being the cache name. The parameter will allow the StatsManager to
store aggregated stats in different caches, which as of now are
different tables. To avoid performing an extra query for table
creation at each insertion/retrieval, a hashmap of existing tables
is kept by the StatsManager itself.

git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntopng@8746 a02cd0c1-8e76-42e1-a119-56f9641475e2