top stats: handle VLANs better
This commit adds better handling of VLANs with respect to viewing
the interface-specific top talkers. In case a host or an AS belongs
to a VLAN, such VLAN is specified to disambiguate potential duplicates.
Also, the VLAN name is hopefully made clearer. Also, in case all hosts
don't belong to any VLAN, no VLAN is displayed in the top talkers
page.

git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntopng@8790 a02cd0c1-8e76-42e1-a119-56f9641475e2