StatsManager: add interface to retrieve real epoch of a minute sampling
Now that the stats manager is based on a full epoch as timestamp, add a
new interface that allows to et the real timestamp for a minute stats.
This is useful, e.g., when we know that we want stats for a certain
minute, but we also want to know if these stats were actually
computed for exactly one minute of traffic or for a bit more or a bit
less seconds. This allows to compute average traffic with higher precision
in case of peaks.

git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntopng@8829 a02cd0c1-8e76-42e1-a119-56f9641475e2