Host.cpp: refresh localhost field of host when finding it by MAC
This commit lets the localhost field of a host be dynamically set
when finding it by MAC. Otherwise each time a host was searched for
by MAC (e.g. for some kind of non-IP packet) the localhost field
was set to true and the host was afterwards treated as local.

git-svn-id: https://svn.ntop.org/svn/ntop/trunk/ntopng@8838 a02cd0c1-8e76-42e1-a119-56f9641475e2